move task-list to shared-module
