+++ src/app/home-page/components/today-tasks/today-tasks.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { TaskService } from '../../services/task.service';
+import { TaskService } from '../../../shared-module/services/task.service';
 import { Task } from '../../types/task';
 
 @Component({
