+++ src/app/shared-module/task-list/task-list.component.scss
